Moved some byos stuff
