+++ frackinship/configs/shipupgrades.config
@@ -1,0 +1,31 @@
+{
+	"crewSize" : {
+		"min" : 0,
+		"totalMin" : 0,
+		"totalMax" : 12
+	},
+	"maxFuel" : {
+		"min" : -10000,
+		"max" : 10000,
+		"totalMin" : 0
+	},
+	"fuelEfficiency" : {
+		"min" : -0.99,
+		"max" : 0.99,
+		"totalMax" : 0.99
+	},
+	"shipSpeed" : {
+		"min" : -100,
+		"max" : 384,
+		"totalMin" : 5
+	},
+	"shipMass" : {
+		"min" : -200,
+		"max" : 200,
+		"totalMin" : 0
+	},
+	"atmosphereSystem" : {
+		"maxPerimeter" : 1000,
+		"requireFilledBackground" : true
+	}
+}
