+++ frackinship/quests/fu_byos.questtemplate
@@ -14,7 +14,7 @@
   ],
   "canBeAbandoned" : false,
   "updateDelta" : 60,
-  "script" : "/quests/scripts/byos/fu_byos.lua",
+  "script" : "/frackinship/scripts/quest/frackinship.lua",
   "scriptConfig" : {
     "portraits" : {
       "default" : "sail"
