+++ interface/ai/fu_byosai.config
@@ -227,48 +227,48 @@
 		"initial" : {
 			"text" : "Ship has sustained major damage when escaping earth. Please select which ship type you want:\n\nFrackin Ship:\nBuild your ship. See tricorder quests when obtained. Requires installation of a new FTL Drive.\n\nVanilla Ship:\nThe default starbound ship for your race.",
 			"buttons" : [
-				{"name" : "Frackin Ship", "newState" : "fuShipSelected"},
+				{"name" : "Frackin Ship", "newState" : "frackinShipSelected"},
 				{"name" : "Vanilla Ship", "newState" : "vanillaShipSelected"}//,
-				//{"name" : "Frackin Ship (Test)", "newState" : "fuShipChoice"}
+				//{"name" : "Frackin Ship (Test)", "newState" : "frackinShipChoice"}
 			],
-			"path" : "root/sail/ui/replaceship"
+			"path" : "root/sail/ui/fixship"
 		},
-		"fuShipChoice" : {
+		"frackinShipChoice" : {
 			"text" : " ",
 			"buttons" : [
-				{"name" : "Select", "newState" : "fuShipSelected"},
-				{"name" : "Preview", "newState" : "fuShipPreview"}
+				{"name" : "Select", "newState" : "frackinShipSelected"},
+				{"name" : "Preview", "newState" : "frackinShipPreview"}
 			],
 			"previousState" : "initial",
-			"path" : "root/sail/ui/replaceship/fu"
+			"path" : "root/sail/ui/fixship/frackin"
 		},
-		"fuShipPreview" : {
+		"frackinShipPreview" : {
 			"text" : " ",
 			"buttons" : [
-				{"name" : "Select", "newState" : "fuShipSelected"}
+				{"name" : "Select", "newState" : "frackinShipSelected"}
 			],
-			"previousState" : "fuShipChoice",
-			"path" : "root/sail/ui/replaceship/fu/preview"
+			"previousState" : "frackinShipChoice",
+			"path" : "root/sail/ui/fixship/frackin/preview"
 		},
-		"fuShipSelected" : {
+		"frackinShipSelected" : {
 			"text" : "The frackin ship. Upgraded differently. This ship type has access to features that the vanilla ship does not have access to.\n\nCurrently Selected Ship: <shipName>\nCurrently Selected Ship Mode: <shipMode>",
 			"buttons" : [
-				{"name" : "Confirm", "newState" : "fuShipChosen"}
+				{"name" : "Confirm", "newState" : "frackinShipChosen"}
 			],
 			"previousState" : "initial",
-			//"previousState" : "fuShipChoice",
-			"path" : "root/sail/ui/replaceship/fu/<shipMode>"
+			//"previousState" : "frackinShipChoice",
+			"path" : "root/sail/ui/fixship/frackin/<shipMode>"
 		},
-		"fuShipChosen" : {
+		"frackinShipChosen" : {
 			"text" : ""
 		},
 		"vanillaShipSelected" : {
-			"text" : "The normal vanilla ship. Upgraded normally. This ship type won't have access to some of the features the FU ship will have access to.",
+			"text" : "The normal vanilla ship. Upgraded normally. This ship type won't have access to some of the features the frackin ship will have access to.",
 			"buttons" : [
 				{"name" : "Confirm", "newState" : "vanillaShipChosen"}
 			],
 			"previousState" : "initial",
-			"path" : "root/sail/ui/replaceship/vanilla"
+			"path" : "root/sail/ui/fixship/vanilla"
 		},
 		"vanillaShipChosen" : {
 			"text" : ""
