--- quests/byos/fu_shipupgrades.questtemplate
@@ -1,24 +1,0 @@
-{
-  "id" : "fu_shipupgrades",
-  "mainQuest" : false,
-  "showInLog" : false,
-  "showAcceptDialog" : false,
-  "showCompleteDialog" : false,
-  "showFailDialog" : false,
-  "title" : "Ship Quest Tracker",
-  "text" : "Adding certain machines to your ship will upgrade its abilities. An FTL drive is required to navigate the stars.",
-  "completionText" : "You shouldn't complete this quest.",
-  "moneyRange" : [0, 0],
-  "rewards" : [
-    [ ]
-  ],
-  "canBeAbandoned" : false,
-  "updateDelta" : 60,
-  "script" : "/quests/scripts/byos/fu_shipupgrades.lua",
-  "scriptConfig" : {
-    "portraits" : {
-      "default" : "sail"
-    },
-	"byosRecipes" : "/interface/ai/fu_byosrecipes.config"
-  }
-}
