+++ quests/story/bootship.questtemplate.patch
@@ -2,6 +2,6 @@
 	{
 		"op" : "replace",
 		"path" : "/script",
-		"value" : "/quests/scripts/story/fu_bootship.lua"
+		"value" : "/frackinship/scripts/quest/fs_bootship.lua"
 	}
 ]
