--- ships/fu_ship.config
@@ -1,84 +1,0 @@
-{
-	"vanillaShip" : {		//uses this as the parameters for player.upgradeShip()
-		"shipLevel" : 1
-	},
-	"racialShips" : {			//REMOVE
-		"shipLevel" : 1,				//the starting level of the racial fu ship
-		"allowOtherRaceShips" : true	//whether the player can use racial ships from other races
-	},
-	"upgradableShips" : [
-		{
-			"type" : "racial",										//will cause the list to be populated by the racial ships of installed races
-			"shipLevel" : 2,										//the starting level of the upgradable ship (NOTE: IS 1 HIGHER THAN VANILLA SHIP UPGRADES)
-			"disallowOtherRaceShips" : false						//optional, whether the player can't use racial ships from other races
-			//"upgradeType" : "racial"								//optional, makes it so that things that trigger upgradable ship upgrades have to be that type to triger it (defaults to racial)
-			//"whitelistedRaces" : {"apex" : true, "avian": true},	//optional, only these racial ships will be options (all racial ships will be options if not set, overrides blacklistedRaces, only works for "type" : "racial")
-			//"blacklistedRaces" : {"floran" : true, "glitch" : true}	//optional, these racial ships will not be options (all racial ships will be options if not set, overridden by whitelistedRaces, only works for "type" : "racial")
-		}
-	],
-	"byosShips" : [
-		{
-			"type" : "fu_byosdefaultship",
-			"name" : "Default BYOS",
-			"offset" : [0, 0],											//optional (defaults to [0,0] if not set)
-			"previewImage" : "/interface/ai/shippreviews/fu_byosdefaultship.png"	//optional
-			//"racialiserOverride" : "human",							//optional, racialises the ship objects to this race(uses the players race if not set)
-			//"raceWhitelist" : {"apex" : true, "avian": true},			//optional, only these races can use the ship (all races can use it if not set, overrides "raceBlacklist")
-			//"raceBlacklist" : {"floran" : true, "glitch" : true},		//optional, these races cannot use the ship (all races can use it if not set, overridden by "raceWhitelist")
-			//"icon" : "/interface/title/sapemale.png"					//optional, the icon used on the ship list
-		},
-		{
-			"type" : "/ships/veluu/veluuT0.structure",
-			"name" : "Structure File BYOS Test (remove)"
-		},
-		{
-			"type" : "BYOS_Apex",
-			"name" : "Apex BYOS Ship",
-			"offset" : [0,0],	//change to align the teleporter to the same spot
-			"racialiserOverride" : "apex",
-			"icon" : "/interface/title/sapemale.png"
-		},
-		{
-			"type" : "BYOS_Avian",
-			"name" : "Avian BYOS Ship",
-			"offset" : [0,0],	//change to align the teleporter to the same spot
-			"racialiserOverride" : "avian",
-			"icon" : "/interface/title/avianmale.png"
-		},
-		{
-			"type" : "BYOS_Floran",
-			"name" : "Floran BYOS Ship",
-			"offset" : [0,0],	//change to align the teleporter to the same spot
-			"racialiserOverride" : "floran",
-			"icon" : "/interface/title/plantmale.png"
-		},
-		{
-			"type" : "BYOS_Glitch",
-			"name" : "Glitch BYOS Ship",
-			"offset" : [0,0],	//change to align the teleporter to the same spot
-			"racialiserOverride" : "glitch",
-			"icon" : "/interface/title/robotmale.png"
-		},
-		{
-			"type" : "BYOS_Human",
-			"name" : "Human BYOS Ship",
-			"offset" : [0,0],	//change to align the teleporter to the same spot
-			"racialiserOverride" : "human",
-			"icon" : "/interface/title/humanmale.png"
-		},
-		{
-			"type" : "BYOS_Hylotl",
-			"name" : "Hylotl BYOS Ship",
-			"offset" : [0,0],	//change to align the teleporter to the same spot
-			"racialiserOverride" : "hylotl",
-			"icon" : "/interface/title/aquaticmale.png"
-		},
-		{
-			"type" : "BYOS_Novakid",
-			"name" : "Novakid BYOS Ship",
-			"offset" : [0,0],	//change to align the teleporter to the same spot
-			"racialiserOverride" : "novakid",
-			"icon" : "/interface/title/novakidmale.png"
-		}
-	]
-}
